ifbench/requirements.txt
dspy-ai
datasets
matplotlib
pandas
